+++ src/Screens/BaseScreen.Canvas.cs
@@ -16,7 +16,7 @@ using CivOne.Graphics.Sprites;
 
 namespace CivOne.Screens
 {
-	public abstract partial class BaseScreen : IDefaultTextSettings
+	public abstract partial class BaseScreen : IDefaultTextSettings, IButtonDrawer
 	{
 		public TextSettings DefaultTextSettings { get; set; }
 
@@ -97,7 +97,7 @@ namespace CivOne.Screens
 				.AddLayer(borders[3], Width - 8, Height - 8);
 		}
 
-		protected void DrawButton(string text, byte fontId, byte colour, byte colourDark, int x, int y, int width, int height)
+		public void DrawButton(string text, byte fontId, byte colour, byte colourDark, int x, int y, int width, int height)
 		{
 			this.FillRectangle(x, y, width, 1, 7)
 				.FillRectangle(x, y + 1, 1, (height - 1), 7)
@@ -107,7 +107,7 @@ namespace CivOne.Screens
 				.DrawText(text, fontId, colourDark, x + (int)Math.Ceiling((double)width / 2), y + 2, TextAlign.Center);
 		}
 
-		protected void DrawButton(string text, byte colour, byte colourDark, int x, int y, int width) => DrawButton(text, 1, colour, colourDark, x, y, width, Resources.GetFontHeight(1) + 3);
+		public void DrawButton(string text, byte colour, byte colourDark, int x, int y, int width) => DrawButton(text, 1, colour, colourDark, x, y, width, Resources.GetFontHeight(1) + 3);
 
 		public void ResetPalette()
 		{